% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{calc_zscore}
\alias{calc_zscore}
\title{calc_zscore}
\usage{
calc_zscore(x, apply_log = TRUE, pseudo = 0.01)
}
\arguments{
\item{pseudo}{}
}
\value{
z score values
}
\description{
calc_zscore
}
\examples{
calc_zscore(runif(100))
}
