% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{data.table_2_matrix}
\alias{data.table_2_matrix}
\alias{matrix_2_data.table}
\title{data.table_to_matrix}
\usage{
data.table_2_matrix(dt)

matrix_2_data.table(mat)
}
\arguments{
\item{dt}{}

\item{mat}{}
}
\value{



}
\description{
create a matrix from a data.table using gene_id as rownmaes
}
\examples{
mat = load_miR_counts()
dt = matrix_2_data.table(mat)
mat2 = data.table_2_matrix(dt)
}
