% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_sample_id_2_patient_id}
\alias{convert_sample_id_2_patient_id}
\title{convert_sample_id_2_patient_id}
\usage{
convert_sample_id_2_patient_id(ids)
}
\arguments{
\item{ids}{extracts patient_id from sample IDs}
}
\value{
TARGET patient ids
}
\description{
convert_sample_id_2_patient_id
}
\examples{
sample_ids = c(
  "TARGET-15-SJMPAL043508-04B-01R",
  "TARGET-15-SJMPAL043511-09B-01R",
  "TARGET-15-SJMPAL043512-03A-01R")
convert_sample_id_2_patient_id(sample_ids)
}
