% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_sample_id_2_sample_code}
\alias{convert_sample_id_2_sample_code}
\title{convert_sample_id_2_sample_code}
\usage{
convert_sample_id_2_sample_code(ids)
}
\arguments{
\item{ids}{}
}
\value{

}
\description{
convert_sample_id_2_sample_code
}
\examples{
sample_ids = c(
  "TARGET-15-SJMPAL043508-04B-01R",
  "TARGET-15-SJMPAL043511-09B-01R",
  "TARGET-15-SJMPAL043512-03A-01R")
convert_sample_id_2_sample_code(sample_ids)
}
