% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append.R
\name{filter_expression_to_valid}
\alias{filter_expression_to_valid}
\title{filter_expression_to_valid}
\usage{
filter_expression_to_valid(dt, clin_dt = load_clinical_data())
}
\arguments{
\item{dt}{matrix or tidy data.table of expression data, as from load_RNA_counts() or convert_matrix_2_data.table()}

\item{clin_dt}{clinical data, as from load_clinical_data()}
}
\value{
Object of same class of dt with samples not covered by clin_dt removed
}
\description{
filter_expression_to_valid
}
\examples{
mat = load_RNA_counts()
clin_dt = load_clinical_data()
mat.valid = filter_expression_to_valid(mat, clin_dt)
}
