% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_wgs_ids}
\alias{load_wgs_ids}
\alias{load_wgs_fusion}
\title{load_wgs_ids}
\usage{
load_wgs_ids()

load_wgs_fusion()
}
\value{
list containing patient ids, sample ids, and data.table of files

data.table of fusion calls from WGS TARGET
}
\description{
load_wgs_ids

load_wgs_fusion
}
\examples{
wgs_desc = load_wgs_ids()
fusion_dt = load_wgs_fusion()
fusion_patients = unique(fusion_dt$patient_id)

seqsetvis::ssvFeatureVenn(list(
  fusion = fusion_patients,
  all = wgs_desc$patient_ids
))

vcf_dt = wgs_desc$vcf_dt
vcf_dt.dbl = vcf_dt[group == "double_sample"]
vcf_dt.dbl[, ids := sub("_TARG", " TARG", ids)]
sample_ids.dbl = unique(unlist(vcf_dt.dbl[, tstrsplit(ids, " ")]))

seqsetvis::ssvFeatureVenn(list(
  double = sample_ids.dbl,
  all = wgs_desc$sample_ids
))

}
