% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runApp.R
\name{plot_meta_alluvial}
\alias{plot_meta_alluvial}
\title{plot_meta_alluvial}
\usage{
plot_meta_alluvial(meta_dt, shown_filters, fill_var = unlist(shown_filters[1]))
}
\arguments{
\item{fill_var}{}
}
\description{
plot_meta_alluvial
}
\examples{
clin_dt = load_clinical_data()
plot_meta_alluvial(clin_dt, shown_filters = c("Phase", "Cell.of.Origin", "Vital.Status"))
}
