% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{plot_survival_discovery_vs_validation}
\alias{plot_survival_discovery_vs_validation}
\title{plot_survival_discovery_vs_validation}
\usage{
plot_survival_discovery_vs_validation(surv_res_dt)
}
\arguments{
\item{surv_res_dt}{}
}
\value{

}
\description{
plot_survival_discovery_vs_validation
}
