% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{plot_survival_goi}
\alias{plot_survival_goi}
\title{plot_survival_goi}
\usage{
plot_survival_goi(
  goi,
  rpm_mat,
  meta_dt,
  title_FUN = function(goi) {
     goi
 }
)
}
\arguments{
\item{goi}{}
}
\description{
plot_survival_goi
}
