% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{run_survival.goi_zscore}
\alias{run_survival.goi_zscore}
\title{run_survival.goi_zscore}
\usage{
run_survival.goi_zscore(
  rpm_mat,
  survival_info,
  goi = "LEF1",
  goi.name = goi,
  pseudo = 0.01,
  breaks = c(-1, 1),
  bin_colors = viridisLite::viridis(length(breaks) + 1),
  return_data_only = FALSE,
  apply_zscore = TRUE
)
}
\arguments{
\item{return_data_only}{if TRUE, omit all plots and only return data}
}
\description{
run_survival.goi_zscore
}
