% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{run_survival}
\alias{run_survival}
\title{run_survival}
\usage{
run_survival(surv_dt, group_var, legend_title = group_var, color = NULL)
}
\arguments{
\item{color}{}
}
\value{

}
\description{
run_survival
}
