% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{run_survival_scan.phaseII}
\alias{run_survival_scan.phaseII}
\title{run_survival_scan.phaseII}
\usage{
run_survival_scan.phaseII(goi_todo, analysis_name, n_todo = length(goi_todo))
}
\arguments{
\item{analysis_name}{}
}
\value{

}
\description{
run_survival_scan.phaseII
}
