% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runApp.R
\name{TARGETdata.runApp.preload}
\alias{TARGETdata.runApp.preload}
\title{TARGETdata.runApp.preload}
\usage{
TARGETdata.runApp.preload(force = FALSE)
}
\arguments{
\item{force}{}
}
\description{
TARGETdata.runApp.preload
}
