% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runApp.R
\name{TARGETdata.runApp}
\alias{TARGETdata.runApp}
\title{TARGETdata.runApp}
\usage{
TARGETdata.runApp()
}
\arguments{
\item{et}{TARGETdata object}
}
\description{
TARGETdata.runApp
}
\examples{
TARGETdata.runApp()
}
