% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_DESeq.R
\name{write_DESeq_results}
\alias{write_DESeq_results}
\title{write_DESeq_results}
\usage{
write_DESeq_results(des, analysis_name, contrast_VAR = "ik_status")
}
\arguments{
\item{contrast_VAR}{}
}
\value{

}
\description{
write_DESeq_results
}
