% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_data.table_2_matrix}
\alias{convert_data.table_2_matrix}
\alias{convert_matrix_2_data.table}
\title{data.table_to_matrix}
\usage{
convert_data.table_2_matrix(dt)

convert_matrix_2_data.table(mat)
}
\arguments{
\item{dt}{}

\item{mat}{}
}
\description{
create a matrix from a data.table using gene_id as rownmaes
}
\examples{
mat = load_miR_counts()
dt = convert_matrix_2_data.table(mat)
mat2 = convert_data.table_2_matrix(dt)
}
