% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{run_survival_scan.phaseII}
\alias{run_survival_scan.phaseII}
\title{run_survival_scan.phaseII}
\usage{
run_survival_scan.phaseII(
  goi_todo,
  rpm_mat,
  analysis_name,
  meta_dt = load_clinical_data(),
  n_todo = length(goi_todo),
  n_cores = getOption("mc.cores", 1),
  title_FUN = function(goi) {
     goi
 }
)
}
\arguments{
\item{goi_todo}{Genes to analyze.  Should be rownames of rpm_mat.}

\item{rpm_mat}{RPM normalized gene expression.}

\item{analysis_name}{Prefix for analysis.}

\item{meta_dt}{result of load_clinical_data()}

\item{n_todo}{Should just be length of goi_todo}

\item{n_cores}{number of cores to use to parallelize processing}

\item{title_FUN}{function to run on gene_id to generate a title.}
}
\value{
data.table of pvalues for Discovery and Validation
}
\description{
run_survival_scan.phaseII
}
\examples{
mat_rpm = load_RNA_RPM()
clin_dt = load_clinical_data()
meta_dt = make_meta_dt(mat_rpm, clin_dt)
meta_dt = meta_dt[Phase \%in\% c("Phase_II_Discovery", "Phase_II_Validation")]
meta_dt[, rnk := frank(runif(.N), ties.method = "first"), .(Phase)]
meta_dt = meta_dt[rnk <= 10]
mat_rpm = filter_expression_to_valid(mat_rpm, meta_dt)
surv_res = run_survival_scan.phaseII(goi_todo = rownames(mat_rpm)[1:5],
    rpm_mat = mat_rpm,
    meta_dt = meta_dt,
    analysis_name = "test",
    n_cores = 1)
}
