% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runApp.R
\name{TARGETdata.runApp}
\alias{TARGETdata.runApp}
\title{TARGETdata.runApp}
\usage{
TARGETdata.runApp()
}
\arguments{
\item{et}{TARGETdata object}
}
\value{

}
\description{
TARGETdata.runApp
}
\examples{
ex_data = system.file("extdata/test_TARGETdata", package = "TARGETdata", mustWork = TRUE)
et = TARGETdata.load(ex_data)
TARGETdata.runApp(et)
}
